Crate biodivine_sketchbook

Modules

algorithms

All the internal algorithms for the inference process and low-level computation details.

app

General functionality for session management, windows, event stack, tab management, and more.

inference

State of the inference analysis and computation solvers.

logging

Custom logging utilities.

sketchbook

State of the BN sketch, and management for most of the editor tab.

Macros

debug

Version of log with DEBUG severity.

error

Version of log with ERROR severity.

generate_property_enums

Generate two enums for property variants - one where each variant contains inner data, and the other (with completely same variants) without the inner data.

info

Version of log with VERBOSE severity.

log

This is an initial attempt at a useful internal log macro. For now, it just logs a message which is prefixed with its severity and location where the message was generated.

warning

Version of log with WARNING severity.