Module biodivine_sketchbook::algorithms::eval_dynamic

Evaluation of dynamic properties (by either HCTL model checker or special procedures).

Modules

_attractors 🔒

Internal algorithms for attractor computation (adapted from AEON). Modified versions of algorithms adapted from AEON. These algorithms can be used to compute attractor states and optimize some procedures.

_trap_spaces

Internal algorithms for trap space computation (adapted from lib-param-bn).

encode

Encode data and template properties into HCTL.

eval

Evaluate all kinds of dynamic properties.

prepare_graph

Prepare graph and symbolic context to handle all dynamic properties.

processed_props

Processed variants of dynamic properties for evaluation.