Module biodivine_sketchbook::algorithms::eval_dynamic::_attractors

Internal algorithms for attractor computation (adapted from AEON). Modified versions of algorithms adapted from AEON. These algorithms can be used to compute attractor states and optimize some procedures.

Modules

itgr 🔒

Interleaved transition guided reduction quickly eliminates most non-attractor states. Interleaved transition guided reduction algorithm.

saturated_reachability 🔒

Reachability algorithms that use saturation for improved efficiency.

xie_beerel 🔒

Xie-Beerel TSCC algorithm

Functions

process_component 🔒

Process a component found by Xie-Beerel (attractor component for a subset of colors). Update the colors_by_num_attrs so that on index i are all colors with i attractors, after taking the new component into account.

sort_colors_by_attr_num

Compute terminal SCCs, and sort all the colors according to how many attractors they have. Returns the vector, where on index i are all colors with i attractors.