Struct biodivine_sketchbook::algorithms::eval_dynamic::_attractors::itgr::Scheduler

struct Scheduler {
    active_variables: Vec<VariableId>,
    universe: GraphColoredVertices,
    processes: Vec<(usize, Box<dyn Process>)>,
    to_discard: Option<GraphColoredVertices>,
}

(internal) Scheduler manages work divided into Processes. It keeps a universe of unprocessed vertices and a list of remaining active variables.

Fields
active_variables: Vec<VariableId>universe: GraphColoredVerticesprocesses: Vec<(usize, Box<dyn Process>)>to_discard: Option<GraphColoredVertices>
Implementations

impl Scheduler

pub fn new( initial: GraphColoredVertices, variables: Vec<VariableId>, ) -> Scheduler

Create a new Scheduler with initial universe and active variables.

pub fn finalize(self) -> (GraphColoredVertices, Vec<VariableId>)

Finalize this scheduler, returning the current universe and active variables.

pub fn discard_variable(&mut self, var: VariableId)

Remove given var from the list of active variables.

pub fn discard_vertices(&mut self, set: &GraphColoredVertices)

Remove given set from the universe of this scheduler.

pub fn spawn<P: 'static + Process>(&mut self, process: P)

Add a new process into this scheduler.

pub fn get_universe(&self) -> &GraphColoredVertices

Get the current universe set of the scheduler.

pub fn get_active_variables(&self) -> &[VariableId]

Get the list of currently active variables.

pub fn is_done(&self) -> bool

True if all processes are finished.

pub fn step(&mut self, graph: &SymbolicAsyncGraph)

If possible, perform one computational step for one of the processes.