Module biodivine_sketchbook::algorithms::eval_static::prepare_graph

Prepare graph and symbolic context to handle all static properties.

Functions

get_fol_extended_symbolic_graph

Prepare the symbolic context and generate the symbolic transition graph for evaluation of FOL formulas. This means we need to prepare symbolic variables to cover all variables in FOL formulas.

prepare_graph_for_static_fol

Prepare the symbolic context and generate the symbolic transition graph for evaluation of the static properties.