Function biodivine_sketchbook::algorithms::eval_static::prepare_graph::prepare_graph_for_static_fol

pub fn prepare_graph_for_static_fol(
    bn: &BooleanNetwork,
    static_props: &Vec<ProcessedStatProp>,
    base_var_name: &str,
    unit: Option<(&Bdd, &SymbolicContext)>,
) -> Result<SymbolicAsyncGraph, String>

Prepare the symbolic context and generate the symbolic transition graph for evaluation of the static properties.

Since all the static properties are encoded as FOL properties, we just need to handle this one case. This means we need to prepare symbolic variables to cover all variables in FOL formulas.

Arg unit is optional unit BDD with potentially different symbolic context (can have different symbolic variables, but has the same bn vars and colors).