Module biodivine_sketchbook::app::state

Modules

_consumed 🔒

_state_app 🔒

_state_atomic 🔒

_state_map 🔒

_undo_stack

editor

Declares top-level state objects that are unique to the sketchbook editor session.

inference

Declares top-level state objects that are unique to the sketchbook inference session.

Structs

AppState

AppState implements mapping between session IDs and session objects.

AtomicState

Atomic state is a SessionState which holds exactly one value of a generic type T.

Enums

Consumed

A Consumed object describes possible outcomes of trying to consume an Event by some session state object.

Traits

SessionHelper

SessionState

StackSession

A Session with a UndoStack with events.

Type Aliases

DynSession

DynSessionState