Module biodivine_sketchbook::app::state::_undo_stack

Structs

UndoStack

The stack that keeps track of all the events that can be reversed.

UndoStackEntry

The items on the undo/redo stack are pairs of events: one performs the actual action while the other reverses said action.

Constants

DEFAULT_EVENT_LIMIT

DEFAULT_PAYLOAD_LIMIT