Module biodivine_sketchbook::inference

State of the inference analysis and computation solvers.

Modules

candidate_sampling 🔒

Utilities to sample and download networks. Some functionality is taken from our repository [biodivine-bn-classifier].

inference_results

Structures and utilities to track final results of inference.

inference_solver

Structures and methods to run the whole inference process. This module covers the general async workflow, the details regarding the actual algorithms are in the module crate::algorithms.

inference_state

Structures and utilities to track the whole state of inference.

inference_status

Structures and utilities to track progress status of inference.

inference_type

Enum with various supported inference types.

results_export 🔒

Utilities to download results.

sampling_data

Struct with details regarding candidate sampling.

update_fn_details 🔒

Utilities to explore canditate update functions.