Module biodivine_sketchbook::sketchbook

State of the BN sketch, and management for most of the editor tab.

Modules

_sketch 🔒

The main Sketch manager object and its utilities.

bn_utils

Utilities regarding conversion of BN components and similar.

data_structs

Structs and utility methods that can be used for communication with frontend.

event_utils 🔒

(internal) Utility functions specifically related to events.

ids

Definitions and utilities for type-safe identifiers of various components.

layout

Structs and utility methods regarding the layout of the Regulations editor.

model

Structs and utility methods regarding the model of the Regulations editor.

observations

Structs and utility methods regarding observations and datasets.

properties

Classes and utility methods regarding properties.

utils 🔒

(internal) General utilities used throughout the module (e.g., serialization helper methods).

Structs

Sketch

Object encompassing all of the individual modules of the Boolean network sketch.

Traits

JsonSerde

Trait that implements to_json_str and from_json_str wrappers to serialize and deserialize objects, utilizing serde_json.

Manager

Trait implementing functionality relevant for all manager structs (such as ModelState, ObservationManager, …).