Module biodivine_sketchbook::sketchbook::data_structs

Structs and utility methods that can be used for communication with frontend.

Modules

_dataset_data 🔒

(internal) Definition and utility methods for DatasetData and DatasetMetaData.

_dynamic_prop_data 🔒

(internal) Definition and utility methods for all kinds of dynamic properties.

_fn_arg_change_data 🔒

(internal) Definition and utility methods for ChangeArgMonotoneData and ChangeArgEssentialData.

_id_change_data 🔒

(internal) Definition and utility methods for ChangeIdData.

_layout_data 🔒

(internal) Definition and utility methods for LayoutData and LayoutMetaData.

_layout_node_data 🔒

(internal) Definition and utility methods for LayoutNodeData.

_model_data 🔒

(internal) Definition and utility methods for ModelData.

_observation_data 🔒

(internal) Definition and utility methods for ObservationData.

_regulation_data 🔒

(internal) Definition and utility methods for RegulationData.

_sketch_data 🔒

(internal) Definition and utility methods for SketchData.

_static_prop_data 🔒

(internal) Definition and utility methods for all kinds of static properties.

_uninterpreted_fn_data 🔒

(internal) Definition and utility methods for UninterpretedFnData.

_variable_data 🔒

(internal) Definition and utility methods for VariableData.

Structs

ChangeArgEssentialData

Structure for receiving data about changes in essentiality of uninterpreted fn’s argument from the frontend.

ChangeArgMonotoneData

Structure for receiving data about changes in monotonicity of uninterpreted fn’s argument from the frontend.

ChangeIdData

Structure for sending data about changes in object’s ID to the frontend.

DatasetData

Structure for sending data about Dataset .

DatasetMetaData

Structure for sending metadata about Dataset. This includes name, id, variable names, but excludes all observations.

DynPropertyData

Structure for sending data about dynamic properties to the frontend.

LayoutData

Structure for sending simplified general data about Layout to the frontend.

LayoutMetaData

Structure for sending metadata about Layout. This includes id, and name, but excludes all actual nodes.

LayoutNodeData

Structure for sending data about NodeLayout to frontend.

LayoutNodeDataPrototype

The same as LayoutNodeData, but does not have a fixed variable ID because it is associated with a variable that does not have an ID yet.

ModelData

Structure for sending/exporting all necessary data about the model part of the sketch.

ObservationData

Structure for sending data about Observation to the frontend.

RegulationData

Structure for sending simplified data about Regulation to the frontend.

SketchData

Structure for sending/exporting data about the whole Sketch.

StatPropertyData

Structure for sending data about static properties to the frontend.

UninterpretedFnData

Structure for sending data about UninterpretedFn to the frontend.

VariableData

Structure for sending data about Variable and its UpdateFn to the frontend.

VariableWithLayoutData

Structure for receiving data about Variable and all of its Layout data from the frontend.

Enums

DynPropertyTypeData

Enum covering all variants of dynamic properties and their necessary data.

StatPropertyTypeData

Enum covering all variants of static properties and their necessary data.