Module biodivine_sketchbook::sketchbook::layout

Structs and utility methods regarding the layout of the Regulations editor.

Modules

_layout 🔒

(internal) Utility methods for Layout.

_node_layout 🔒

(internal) Utility methods for NodeLayout.

_node_position 🔒

(internal) Utility methods for NodePosition.

Structs

Layout

Structure to capture all the layout data regarding one particular layout of the regulations editor.

LayoutNode

A node for a particular Variable’s in a particular Layout. This structure only holds information about the node itself, the id of the variable and layout is held elsewhere (see super::Layout or crate::sketchbook::model::ModelState).

NodePosition

Position of a particular LayoutNode at the editor’s layout. Coordinates are floating point and can be negative.

Type Aliases

LayoutNodeIterator

An iterator over all (VarId, NodeLayout) pairs of a Layout.