Struct biodivine_sketchbook::sketchbook::layout::_layout::Layout

pub struct Layout {
    name: String,
    nodes: HashMap<VarId, LayoutNode>,
}

Structure to capture all the layout data regarding one particular layout of the regulations editor.

Fields
name: Stringnodes: HashMap<VarId, LayoutNode>
Implementations

impl Layout

Methods for safely constructing or mutating instances of Layout.

pub fn new_empty(name: &str) -> Result<Layout, String>

Create new empty Layout (i.e., with no nodes) with a given name.

pub fn new( name: &str, var_node_pairs: Vec<(&str, LayoutNode)>, ) -> Result<Layout, String>

Create new Layout with a given name and nodes.

pub fn new_from_another_copy(name: &str, template_layout: &Layout) -> Layout

Create new Layout with a given name, that is a direct copy of another existing valid template_layout.

pub fn new_from_vars_default( name: &str, variables: Vec<VarId>, ) -> Result<Layout, String>

Create new Layout with a given name, which will contain nodes all the given variables, all of the nodes will be located at a default position.

Returns Error if given ids contain duplicates.

pub fn set_layout_name(&mut self, name: &str) -> Result<(), String>

Rename this Layout.

pub fn add_node(&mut self, var: VarId, node: LayoutNode) -> Result<(), String>

Add a new (pre-generated) node.

You must ensure that the variable is valid before adding it to the layout.

Returns Err if there already is a node for this variable.

pub fn add_node_by_coords( &mut self, var: VarId, p_x: f32, p_y: f32, ) -> Result<(), String>

Add a new node for a given variable to this layout. Currently, nodes only hold information regarding a 2D position.

You must ensure that the variable is valid before adding it to the layout.

Returns Err if there already is a node for this variable.

pub fn add_default_node(&mut self, variable: VarId) -> Result<(), String>

Add a new default node (at 0,0) for a given variable.

You must ensure that the variable is valid before adding it to the layout.

Returns Err if there already is a node for this variable.

pub fn update_node_position( &mut self, variable: &VarId, new_x: f32, new_y: f32, ) -> Result<(), String>

Update position of a node for a given variable.

Return Err if variable did not have a corresponding node in this layout.

pub fn remove_node(&mut self, variable: &VarId) -> Result<(), String>

Remove a node for a given variable from this layout.

Return Err if variable did not have a corresponding node in this layout.

pub fn change_node_id( &mut self, original_id: &VarId, new_id: VarId, ) -> Result<(), String>

Change id of a variable with original_id to new_id.

impl Layout

Utility methods to assert (non-)existence of nodes in the layout.

fn assert_no_variable(&self, var_id: &VarId) -> Result<(), String>

(internal) Utility method to ensure there is no node for the variable with given Id yet.

fn assert_valid_variable(&self, var_id: &VarId) -> Result<(), String>

(internal) Utility method to ensure there is a node for a variable with given Id.

impl Layout

Methods for observing instances of ModelState (various getters, etc.).

pub fn get_node(&self, variable: &VarId) -> Result<&LayoutNode, String>

Layout information regarding the node for a particular variable.

pub fn get_layout_name(&self) -> &String

Human-readable name of this layout.

pub fn get_node_position( &self, variable: &VarId, ) -> Result<&NodePosition, String>

Number of nodes in this layout.

pub fn get_num_nodes(&self) -> usize

Number of nodes in this layout.

pub fn layout_nodes(&self) -> LayoutNodeIterator<'_>

Return an iterator over all nodes of this layout.