Module biodivine_sketchbook::sketchbook::model

Structs and utility methods regarding the model of the Regulations editor.

Modules

_binary_op 🔒

(internal) Definition and utility methods for BinaryOp.

_essentiality 🔒

(internal) Definition and utility methods for Essentiality.

_function_tree 🔒

(internal) Definition and utility methods for FnTree.

_model_state 🔒

(internal) Definition and utility methods for the manager class ModelState.

_monotonicity 🔒

(internal) Definition and utility methods for Monotonicity.

_regulation 🔒

(internal) Definition and utility methods for Regulation.

_uninterpreted_fn 🔒

(internal) Definition and utility methods for UninterpretedFn.

_uninterpreted_fn_arg 🔒

(internal) Definition and utility methods for FnArgument.

_update_function 🔒

(internal) Definition and utility methods for UpdateFn.

_variable 🔒

(internal) Definition and utility methods for Variable.

Structs

FnArgument

Data regarding an argument of an uninterpreted function.

ModelState

Object representing the state of the model in the Boolean network editor. The model encompasses variables, regulations, uninterpreted functions, update functions, and layout information.

Regulation

Describes an interaction between two variables, regulator and target. Every regulation can be monotonous and can be set as essential:

UninterpretedFn

An uninterpreted function of a partially specified model.

UpdateFn

Update function of a BooleanNetwork.

Variable

A type safe object for a Boolean variable of a ModelState.

Enums

BinaryOp

Enum for all possible Boolean binary operators occurring in update functions or first-order formulas (for static properties).

Essentiality

Possible variants of essentiality of a Regulation.

FnTree

Syntactic tree of a partially defined Boolean function. This might specify an update function, or a partially defined uninterpreted fn.

Monotonicity

Possible variants of (non)-monotonous effects of a Regulation.

Type Aliases

LayoutIterator

An iterator over all (LayoutId, Layout) pairs of a ModelState.

RegulationIterator

An iterator over all Regulations of a ModelState.

UninterpretedFnIterator

An iterator over all (UninterpretedFnId, UninterpretedFn) pairs of a ModelState.

UpdateFnIterator

An iterator over all (VarId, UpdateFn) pairs of a ModelState.

VariableIterator

An iterator over all (VarId, Variable) pairs of a ModelState.