Module biodivine_sketchbook::sketchbook::model::_model_state

(internal) Definition and utility methods for the manager class ModelState.

Modules

_impl_convert_bn 🔒

(internal) Methods for converting between ModelState and BooleanNetwork.

_impl_convert_reg_graph 🔒

(internal) Methods for converting between ModelState and RegulatoryGraph.

_impl_editing 🔒

(internal) Methods for safely constructing or editing instances of ModelState.

_impl_id_generating 🔒

(internal) Implementation of the safe identifier generating.

_impl_observing 🔒

(internal) Methods for observing instances of ModelState (various getters, etc.).

_impl_session_state 🔒

(internal) Implementation of event-based API for the crate::app::state::SessionState trait.

Structs

ModelState

Object representing the state of the model in the Boolean network editor. The model encompasses variables, regulations, uninterpreted functions, update functions, and layout information.