Module biodivine_sketchbook::sketchbook::properties::static_props::_property_types

(internal) Variants of static properties.

Structs

FnInputEssential

Variant of StatProperty requiring that an input (on specified input_index) of a target function is essential - either generally, or within optionally specified context.

FnInputMonotonic

Variant of StatProperty requiring that an input (on specified input_index) of a target function is monotonic - either generally, or within optionally specified context.

GenericStatProp

Variant of StatProperty requiring that a particular first-order formula is satisfied. The struct carries the user-provided raw_formula string, as well as its processed internal version processed_formula.

RegulationEssential

Variant of StatProperty requiring that a regulation input -> target is essential, either generally, or within optionally specified context.

RegulationMonotonic

Variant of StatProperty requiring that a regulation input -> target is monotonic, either generally, or within optionally specified context.

Enums

SimpleStatPropertyType

Enum covering all variants of static properties (only discriminants, no data).

StatPropertyType

Enum covering all variants of static properties and their necessary data.

Functions

are_same_stat_variant

Check if two StatPropertyType instances are of the same variant.