Struct biodivine_sketchbook::sketchbook::properties::static_props::_static_property::StatProperty

pub struct StatProperty {
    name: String,
    annotation: String,
    variant: StatPropertyType,
}

A typesafe representation of various kinds of static properties. Each property has a name and field variant encompassing inner data.

The formula that will be internally created (usually, apart from generic variant) depends on particular type of the property - there are multiple variants of properties, each carrying its own different metadata that are later used to build the formula.

Fields
name: Stringannotation: Stringvariant: StatPropertyType
Implementations

impl StatProperty

Creating static properties.

fn new_raw( name: &str, variant: StatPropertyType, annotation: &str, ) -> StatProperty

(internal) Shorthand to create a property given its already created internal StatPropertyType data, name, and annotation.

pub fn try_mk_generic( name: &str, raw_formula: &str, annotation: &str, ) -> Result<StatProperty, String>

Create “generic” StatProperty instance directly from a formula, which must be in a correct format (which is verified).

pub fn mk_regulation_essential( name: &str, input: Option<VarId>, target: Option<VarId>, value: Essentiality, annotation: &str, ) -> StatProperty

Create StatProperty instance describing that an input of an update function is essential.

pub fn mk_regulation_essential_context( name: &str, input: Option<VarId>, target: Option<VarId>, value: Essentiality, context: String, annotation: &str, ) -> StatProperty

Create StatProperty instance describing that an input of an update function is essential in a certain context.

pub fn mk_regulation_monotonic( name: &str, input: Option<VarId>, target: Option<VarId>, value: Monotonicity, annotation: &str, ) -> StatProperty

Create StatProperty instance describing that an input of an update function is monotonic.

pub fn mk_regulation_monotonic_context( name: &str, input: Option<VarId>, target: Option<VarId>, value: Monotonicity, context: String, annotation: &str, ) -> StatProperty

Create StatProperty instance describing that an input of an update function is monotonic in a certain context.

pub fn mk_fn_input_essential( name: &str, input_index: Option<usize>, target: Option<UninterpretedFnId>, value: Essentiality, annotation: &str, ) -> StatProperty

Create StatProperty instance describing that an input of an uninterpreted function is essential.

pub fn mk_fn_input_essential_context( name: &str, input_index: Option<usize>, target: Option<UninterpretedFnId>, value: Essentiality, context: String, annotation: &str, ) -> StatProperty

Create StatProperty instance describing that an input of an uninterpreted function is essential in a certain context.

pub fn mk_fn_input_monotonic( name: &str, input_index: Option<usize>, target: Option<UninterpretedFnId>, value: Monotonicity, annotation: &str, ) -> StatProperty

Create StatProperty instance describing that an input of an uninterpreted function is monotonic.

pub fn mk_fn_input_monotonic_context( name: &str, input_index: Option<usize>, target: Option<UninterpretedFnId>, value: Monotonicity, context: String, annotation: &str, ) -> StatProperty

Create StatProperty instance describing that an input of an uninterpreted function is monotonic in a certain context.

pub fn default(variant: SimpleStatPropertyType) -> StatProperty

Create default StatProperty instance of specified variant.

pub fn default_generic() -> StatProperty

Create default “generic” StatProperty instance, representing “true” formula.

pub fn default_regulation_essential() -> StatProperty

Create default StatProperty instance for regulation essentiality (with empty input and target fields and Unknown essentiality).

pub fn default_regulation_essential_context() -> StatProperty

Create default StatProperty instance for regulation essentiality in a context (with empty input, target, and context fields and Unknown essentiality).

pub fn default_regulation_monotonic() -> StatProperty

Create default StatProperty instance for regulation monotonicity (with empty input and target fields and Unknown monotonicity).

pub fn default_regulation_monotonic_context() -> StatProperty

Create default StatProperty instance for regulation monotonicity in a context (with empty input, target, and context fields and Unknown monotonicity).

pub fn default_fn_input_essential() -> StatProperty

Create default StatProperty instance for function input essentiality (with empty input and target fields and Unknown essentiality).

pub fn default_fn_input_essential_context() -> StatProperty

Create default StatProperty instance for function input essentiality in a context (with empty input, target, and context fields and Unknown essentiality).

pub fn default_fn_input_monotonic() -> StatProperty

Create default StatProperty instance for function input monotonicity (with empty input and target fields and Unknown monotonicity).

pub fn default_fn_input_monotonic_context() -> StatProperty

Create default StatProperty instance for function input monotonicity in a context (with empty input, target, and context fields and Unknown monotonicity).

impl StatProperty

Editing static properties.

pub fn set_name(&mut self, new_name: &str) -> Result<(), String>

Set property’s name.

pub fn set_annotation(&mut self, annotation: &str)

Set property’s annotation string.

pub fn set_input_var(&mut self, new_var: VarId) -> Result<(), String>

Update property’s sub-field for input variable (of an update fn), where applicable. If not applicable, return Err.

pub fn set_input_index(&mut self, new_idx: usize) -> Result<(), String>

Update property’s sub-field for index of input (of an uninterpreted fn), where applicable. If not applicable, return Err.

pub fn set_target_fn( &mut self, new_target: UninterpretedFnId, ) -> Result<(), String>

Update property’s sub-field for target uninterpreted fn, where applicable. If not applicable, return Err.

pub fn set_target_var(&mut self, new_target: VarId) -> Result<(), String>

Update property’s sub-field for target variable, where applicable. If not applicable, return Err.

pub fn set_monotonicity( &mut self, monotonicity: Monotonicity, ) -> Result<(), String>

Update property’s sub-field for monotonicity, where applicable. If not applicable, return Err.

pub fn set_essentiality( &mut self, essentiality: Essentiality, ) -> Result<(), String>

Update property’s sub-field for essentiality, where applicable. If not applicable, return Err.

pub fn set_context(&mut self, context: String) -> Result<(), String>

Update property’s sub-field for context, where applicable. If not applicable, return Err.

pub fn set_formula(&mut self, new_formula: &str) -> Result<(), String>

Update generic property’s formula. If not applicable (different variant), return Err.

pub fn set_var_id_if_present( &mut self, old_id: VarId, new_id: VarId, ) -> Result<(), String>

If the property is referencing the given variable (as either regulator or target), set that variable to the new value.

If not applicable, return Err.

impl StatProperty

Observing static properties.

pub fn get_name(&self) -> &str

Get property’s name.

pub fn get_annotation(&self) -> &str

Get annotation string.

pub fn get_prop_data(&self) -> &StatPropertyType

Get property’s variant with all the underlying data.

pub fn assert_fully_filled(&self) -> Result<(), String>

Check that the property has all required fields filled out. If some of the required field is set to None, return error.

pub fn get_regulator_and_target( &mut self, ) -> Result<(Option<VarId>, Option<VarId>), String>

Get property’s sub-fields for regulator variable and target variable, where applicable. If not applicable, return Err.

impl StatProperty

Static methods to automatically generate IDs to encode regulation properties.

pub fn get_monotonicity_prop_id( regulator: &VarId, target: &VarId, ) -> StatPropertyId

Get ID of a static property that describes monotonicity of a regulation between regulator and target.

pub fn get_essentiality_prop_id( regulator: &VarId, target: &VarId, ) -> StatPropertyId

Get ID of a static property that describes essentiality of a regulation between regulator and target.